refactor cell sets dict functions
